Memory leak may happen, when calculated min thread.
There may be other application using the dll - min thread is not the no. of CPUs, as the code means.
Whether min no. of threads is big (>100 or even less), there may be memory leak.

The no. of CPUs is more correct, and not causing memory leak.
Hence, change the code of : ThreadPool.GetMinThread( ...) to:
workerThreads = Environment.ProcessorCount;
completionPortThreads = Environment.ProcessorCount;